Add API routes and tests.

+++ src/repositories/sqlalchemy/unit_of_work_sqlalchemy.py
@@ -2,8 +2,13 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import sessionmaker
 
 from src.domain.exceptions import UrlNotFoundException, ShortCodeAlreadyExists
+from src.infrastructure.database import session_maker
 from src.repositories.sqlalchemy.url_repository_sqlalchemy import UrlRepository
 from src.repositories.unit_of_work import UnitOfWorkAbstract
+
+
+def create_uow() -> UnitOfWorkAbstract:
+    return UnitOfWorkSqlAlchemy(session_maker=session_maker)
 
 
 class UnitOfWorkSqlAlchemy(UnitOfWorkAbstract):
@@ -28,9 +33,9 @@ class UnitOfWorkSqlAlchemy(UnitOfWorkAbstract):
             if not str(e).startswith("CHECK constraint failed: LENGTH"):
                 raise ShortCodeAlreadyExists() from e
         except Exception as e:
-            raise Exception from e
-        finally:
+            print(e)
             self.rollback()
+            raise Exception from e
 
     def rollback(self):
         self.session.rollback()
@@ -45,4 +50,8 @@ class UnitOfWorkSqlAlchemy(UnitOfWorkAbstract):
         return url
 
     def delete_by_short_code(self, short_code: str) -> None:
-        self.url_repo.delete_url(short_code=short_code)
+        url = self.url_repo.get_url(short_code=short_code)
+        if url:
+            self.url_repo.delete_url(short_code=short_code)
+        else:
+            raise UrlNotFoundException(short_code=short_code)
